[crawler] refactor play_block code
